+++ CHANGELOG.md
@@ -13,6 +13,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - User can provide the new academy urn once it has be created in Get information
   about schools
 - Added a new list of all projects with Academy URNs
+- Display local authorities and their details
+- Allow local authority records to be edited
+- Allow new local authority records to be added
+- Allow local authority records to be deleted
+- Ensure only the new `service_support` user can create, edit and delete local
+  authority records
 
 ### Changed